A-Packages: Add commands and tasks package
