+++ src/main/java/chatbox/main/commands/Command.java
@@ -1,4 +1,9 @@
-package chatbox.main;
+package chatbox.main.commands;
+
+import chatbox.main.ChatBoxException;
+import chatbox.main.Storage;
+import chatbox.main.tasks.TaskList;
+import chatbox.main.Ui;
 
 public abstract class Command {
     public abstract void execute(TaskList tasks, Ui ui, Storage storage) throws ChatBoxException;
