+++ src/main/java/chatbox/main/tasks/Event.java
@@ -1,5 +1,7 @@
-package chatbox.main;
+package chatbox.main.tasks;
 //Inheritance of Task
+import chatbox.main.DateParser;
+
 import java.time.LocalDateTime;
 public class Event extends Task {
     protected LocalDateTime from;
@@ -12,6 +14,13 @@ public class Event extends Task {
         if (this.from == null || this.to == null) {
             throw new IllegalArgumentException("You key in the format that I don't understand! Please Use d/M/yyyy HHmm (e.g., 2/12/2019 1800)");
         }
+    }
+    public java.time.LocalDateTime getFrom() {
+        return from;
+    }
+
+    public java.time.LocalDateTime getTo() {
+        return to;
     }
     @Override
     public String toString() {
